requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
python-jose[cryptography]
httpx
python-dotenv
groq